Wed, Aug 31, 2022, 4:30 AM +00:00

+++ filter-less/helpers.c
@@ -7,9 +7,9 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     // nested for loop to iterate through each pixel of the image
     for (int i = 0; i < height; i++) {
         for (int j = 0; j < width; j++) {
-            int red = image[i][j].rgbtRed;
-            int green = image[i][j].rgbtGreen;
-            int blue = image[i][j].rgbtBlue;
+            float red = image[i][j].rgbtRed;
+            float green = image[i][j].rgbtGreen;
+            float blue = image[i][j].rgbtBlue;
 
             int average = (red + green + blue) / sizeof(RGBTRIPLE);
